requirements.txt
asgiref==3.2.3  
Django==3.0.3
mysqlclient==1.4.6  
Pillow==7.0.0  
pytz==2019.3 
sqlparse==0.3.0  
wheel==0.33.6 
django-swingtime==1.1.0
django-mysql==3.3.0
django-scheduler==0.9.3
djangorestframework==3.10.03
django-crispy-forms==1.8.1
PyJWT==1.7.1
django-markdown-deux==1.0.5
django-pagedown==2.0.3
markdown2==2.3.8
djangorestframework-jwt==1.11.0
django-debug-toolbar
django-material
